# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "description": "Documentation officielle de l'API Shop",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/products/store": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Créer un produit",
                "operationId": "90841008e6d2f8cb3655547d2377fb01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "nom",
                                    "description",
                                    "stock",
                                    "prix_unitaire",
                                    "image"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "PC"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "PC authentique"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "prix_unitaire": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "prix_promo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 90000
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit créé avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/update/{product}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Modifier un produit",
                "operationId": "40426ac61ca1ce2d315838f9b9163ec9",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "nom",
                                    "description",
                                    "stock",
                                    "prix_unitaire"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "PC"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "PC authentique"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "prix_unitaire": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "prix_promo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 90000
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit modifié avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/destroy/{product}": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Supprimer un produit",
                "operationId": "4fc70012f13b672706315f4f12c1222f",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produit supprimé avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}